Update index.md
Hi Matt, it seems like adding the Master GraphQL Query to the bottom of this page is the most logical location.  I double-checked the query to make sure there were no validation errors.  Please review the description below the section title and also the query itself.   The 4 report numbers in the query all point to various sample reports.  Thanks